da/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
ca-certificates=2021.9.30=hecd8cb5_1
certifi=2021.5.30=py37hecd8cb5_0
cffi=1.15.0=pypi_0
libcxx=12.0.0=h2f01273_0
libffi=3.3=hb1e8313_2
ncurses=6.2=h0a44026_1
numpy=1.21.2=pypi_0
openssl=1.1.1l=h9ed2024_0
pandas=1.3.3=pypi_0
pip=21.0.1=py37hecd8cb5_0
pycparser=2.20=pypi_0
pynacl=1.4.0=pypi_0
python=3.7.10=h88f2d9e_0
python-dateutil=2.8.2=pypi_0
pytz=2021.3=pypi_0
readline=8.1=h9ed2024_0
setuptools=58.0.4=py37hecd8cb5_0
six=1.16.0=pypi_0
sqlite=3.36.0=hce871da_0
tk=8.6.11=h7bc2e8c_0
wheel=0.37.0=pyhd3eb1b0_1
xz=5.2.5=h1de35cc_0
zlib=1.2.11=h1de35cc_3
